+++ src/api/extras.ts
@@ -1,9 +1,6 @@
 import {IDepTreeNode} from "../core/observable";
-import {ComputedValue} from "../core/computedvalue";
-import {Reaction}from "../core/reaction";
-import {unique, invariant} from "../utils/utils";
+import {unique} from "../utils/utils";
 import {getAtom} from "../types/type-utils";
-import {globalState} from "../core/globalstate";
 
 export interface IDependencyTree {
 	name: string;
@@ -39,26 +36,4 @@ function nodeToObserverTree(node: IDepTreeNode): IObserverTree {
 	if (node.observers && node.observers.length)
 		result.observers = <any>unique(node.observers).map(<any>nodeToObserverTree);
 	return result;
-}
-
-export function whyRun(thing?: any, prop?: string) {
-	switch (arguments.length) {
-		case 0:
-			thing = globalState.derivationStack[globalState.derivationStack.length - 1];
-			if (!thing) {
-				console.log("whyRun() can only be used if a derivation is active, or by passing an computed value / reaction explicitly.");
-				return;
-			}
-			break;
-		case 2:
-			thing = getAtom(thing, prop);
-			break;
-	}
-	thing = getAtom(thing);
-	if (thing instanceof ComputedValue)
-		console.log(thing.whyRun());
-	else if (thing instanceof Reaction)
-		console.log(thing.whyRun());
-	else
-		invariant(false, "whyRun can only be used on reactions and computed values");
 }